components design set up wip
