+++ Base.module.css
@@ -1,0 +1,3 @@
+/* .body{
+  display: flex;
+} */
