+++ Base.tsx
@@ -1,17 +1,19 @@
 import React from 'react'
 import styles from "./Base.module.css";
 import NewsTimeLine from './components/NewsTimeLine';
-import * as NewsType from "@/types/News";
+// import * as NewsType from "@/types/News";
+
+import 'normalize.css';
 
 const Base = () => {
-  const newsData: NewsType.News[] = [ /*...*/ ];
+  // const newsData: NewsType.News[] = [ /*...*/ ];
 
   return (
     <div className={styles.body}>
       <div>
         Hello from base
       </div>
-      <NewsTimeLine news={newsData} />
+      <NewsTimeLine />
     </div>
   )
 }
