+++ components/Nav.module.css
@@ -1,0 +1,8 @@
+/* #2BC859
+#FEC005 */
+
+.body{
+  background-color: white;
+  color: black;
+  padding: 10px 20px;
+}
