+++ components/NewsTimeLine.module.css
@@ -1,0 +1,3 @@
+.body{
+  background-color: grey;
+}
