+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "axios": "^1.3.4",
     "next": "latest",
+    "normalize.css": "^8.0.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0"
   },
