+++ styles/Home.module.css
@@ -1,0 +1,9 @@
+.body{
+  display: flex;
+}
+
+@media (max-width: 720px) {
+  .body{
+    flex-direction: column;
+  }
+}
